requirements.txt
streamlit==1.31.0
scikit-learn==1.4.0
openai==1.12.0
python-dotenv==1.0.0
numpy~=1.26.4